+++ lib/src/colors.dart
@@ -1,0 +1,40 @@
+/// {@nodoc}
+final String reset = '\x1B[0m';
+
+/// {@nodoc}
+class ForeColor {
+  static final String black = '\x1B[30m';
+
+  static final String red = '\x1B[31m';
+
+  static final String green = '\x1B[32m';
+
+  static final String yellow = '\x1B[33m';
+
+  static final String blue = '\x1B[34m';
+
+  static final String cyan = '\x1B[36m';
+
+  static final String white = '\x1B[37m';
+
+  static final String orange = '\x1B[38;5;202m';
+}
+
+/// {@nodoc}
+class BackColor {
+  static final String black = "\x1B[40m";
+
+  static final String red = "\x1B[41m";
+
+  static final String green = "\x1B[42m";
+
+  static final String yellow = "\x1B[43m";
+
+  static final String blue = "\x1B[44m";
+
+  static final String purple = "\x1B[45m";
+
+  static final String cyan = "\x1B[46m";
+
+  static final String white = "\x1B[47m";
+}