+++ lib/tarsier_logger.dart
@@ -1,5 +1,3 @@
-/// Tarsier Logger Library
-///
 /// This library provides a simple and flexible logging utility for Dart and Flutter applications.
 /// It supports multiple log levels (info, success, warning, error) and customizable log outputs.
 /// Optionally, emoji icons can be included in log messages for better readability.
@@ -11,5 +9,6 @@
 library tarsier_logger;
 
 // Exporting library components
-export 'src/constants.dart'; // Provides constants for color codes and emoji mapping.
 export 'src/tarsier_logger_base.dart'; // Contains the core TarsierLogger class implementation.
+export 'src/prints.dart';
+export 'src/console.dart';